Remove old converters, remove faulty overloads
