--- src/Core/BuildTools/Converters/VKXmlConvert/VKXmlConvert.cs
@@ -1,12 +1,0 @@
-// This file is part of Silk.NET.
-// 
-// You may modify and distribute Silk.NET under the terms
-// of the MIT license. See the LICENSE file for details.
-
-namespace Silk.NET.BuildTools.VKXmlConvert
-{
-    public class VKXmlConvert
-    {
-        
-    }
-}
