+++ src/OpenGL/Silk.NET.OpenGLES/Silk.NET.OpenGLES.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <LangVersion>latest</LangVersion>
+    <LangVersion>8</LangVersion>
   </PropertyGroup>
 
   <Import Project="..\..\..\build\props\bindings.props" />
